+++ uni_modules/uv-rate/components/uv-rate/uv-rate.vue
@@ -232,7 +232,7 @@
 				setTimeout(() => {
 					this.moving = false;
 				}, 10);
-			},
+			}
 		}
 	}
 </script>